backend/requirements.txt
# FastAPI核心
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# 数据库
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Redis
redis==5.0.1
hiredis==2.3.2

# 认证
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# AI和OCR
openai==1.10.0
pillow==10.2.0
tencentcloud-sdk-python==3.0.1090

# 数据处理
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP客户端
httpx==0.26.0
aiohttp==3.9.1

# 工具
python-dateutil==2.8.2
pytz==2024.1

# 数据导出
openpyxl==3.1.2
